move portable text components to its file
